requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
chromadb
pypdf
python-dotenv
streamlit
sentence-transformers
pysqlite3-binary
duckduckgo-search
